src/testapi/requirements.txt
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.24.0
requests>=2.28.0,<2.33.0
pytest>=7.0.0,<8.4.0
httpx==0.25.1
pyjwt>=2.6.0,<2.8.0
email-validator>=2.0.0,<2.2.0
python-jose==3.3.0
python-multipart==0.0.6
